Popup glosses using `floating-ui`
This will likely need to be more sophisticated; using any block-level markup in the glosses borks the markdown markup, so having the popup elements out-of-tree might make more sense (can deduplicate (and possibly pre-render?) that way too).

This will do as the basis for further thought and discussion, at least.